# frontend/src/shared/api/openapi.json
{
  "components": {
    "schemas": {
...
      "TenantResponse": {
        "description": "Tenant identity payload for the protected example route.",
        "properties": {
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "TenantResponse",
        "type": "object"
      },
...
      "UserMeResponse": {
        "description": "Public current-user payload (read-only, no secrets).",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_superuser": {
            "title": "Is Superuser",
            "type": "boolean"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "plan",
          "is_verified",
          "is_superuser"
        ],
        "title": "UserMeResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "fastapiusersauth",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Foresignal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/users/me": {
      "get": {
        "description": "Return the authenticated user's profile.\n\nBehind `current_user` dependency: returns 401 without a valid auth cookie.\nTenant-scoped: only reads data belonging to the requesting user.",
        "operationId": "get_users_me_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMeResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Users Me",
        "tags": [
          "users"
        ]
      }
    },
...
    "/v1/users/me/tenant": {
      "get": {
        "description": "Protected example route (AC2): 401 without a token, tenant id with one.",
        "operationId": "read_my_tenant_v1_users_me_tenant_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Read My Tenant"
      }
    },
...
  }
}